% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageArea.R
\name{areaColoured}
\alias{areaColoured}
\title{Calculate the area of an image matching a specific colour}
\usage{
areaColoured(file, colour = c(0, 0, 0))
}
\arguments{
\item{file}{A jpeg file to analyse}

\item{colour}{Numeric vector of colour components, c( red, green, blue )}
}
\value{
A numeric value, pixel count of matching colour
}
\description{
Calculate the area of an image matching a specific colour
}
\keyword{analysis}
\keyword{image}
\keyword{jpeg}
