% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageArea.R
\name{areaThreshold}
\alias{areaThreshold}
\title{Calculate the area of an image with colour EXCEEDING threshold grey value}
\usage{
areaThreshold(file, thresholdGrey = 0.5, which = "high")
}
\arguments{
\item{file}{A jpeg file to analyse}

\item{thresholdGrey}{Numeric value 0-1, threshold greyscale value}

\item{which}{Either "high" or "low", specifying what type of thresholding to perform}
}
\value{
A numeric value, pixel count of matching colour
}
\description{
Calculate the area of an image with colour EXCEEDING threshold grey value
}
\keyword{analysis}
\keyword{image}
\keyword{jpeg}
