% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoLevels.R
\name{autoLevels}
\alias{autoLevels}
\title{calculate and apply a linear levels function to an image}
\usage{
autoLevels(image, clipBlacks = 0.02, clipWhites = 0.02)
}
\arguments{
\item{image}{An array, imported from an image file.}

\item{clipBlacks}{Numeric value 0-1. Fraction of image area to clip to black.}

\item{clipWhites}{Numeric value 0-1. Fraction of image area to clip to white.}
}
\description{
calculate and apply a linear levels function to an image
}
\keyword{image}
\keyword{levels}
\keyword{modify}
