% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blankRow.R
\name{blankRow}
\alias{blankRow}
\title{A function for use when adding a data row to a given data frame}
\usage{
blankRow(x, n = 1L)
}
\arguments{
\item{x}{a data.frame or data.table to have a row added}
}
\description{
A function for use when adding a data row to a given data frame
}
