% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentImage.R
\name{cropAndOutput}
\alias{cropAndOutput}
\title{Functions for segmenting an image into a grid of smaller images}
\usage{
cropAndOutput(input.image, grid.coord.upperleft.x, grid.coord.upperleft.y,
  image.width, image.height, output.filename)
}
\arguments{
\item{input.image}{file path to input file}

\item{grid.coord.upperleft.x}{pixel coordinate value}

\item{grid.coord.upperleft.y}{pixel coordinate value}

\item{image.width}{pixels}

\item{image.height}{pixels}

\item{output.filename}{file path to output file}
}
\description{
Functions for segmenting an image into a grid of smaller images
}

