% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENtoGPS.R
\name{ENtoGPS}
\alias{ENtoGPS}
\title{Converting easting and northing values to GPS coordinates}
\usage{
ENtoGPS(eastings, northings, input.projection = "+init=epsg:28355")
}
\arguments{
\item{eastings}{numeric vector}

\item{northings}{numeric vector}

\item{input.projection}{character string specifying projection model of inputs}
}
\description{
Converting easting and northing values to GPS coordinates
}
\keyword{gps}
\keyword{spatial}
