% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exifRead.R
\name{exifRead}
\alias{exifRead}
\title{Extract filename, directory, date, and time from image EXIF data.}
\usage{
exifRead(paths, chunk.size = 500)
}
\arguments{
\item{paths}{A vector of file or folder paths to be analysed}

\item{chunk.size}{How many files to process at a time}
}
\description{
Extract filename, directory, date, and time from image EXIF data.
}
\keyword{date}
\keyword{exif}
\keyword{image}
\keyword{name}
\keyword{time}
