% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exifRead_folder.R
\name{exifRead_folder}
\alias{exifRead_folder}
\title{Read all exif data from a directory of image files recursively.}
\usage{
exifRead_folder(dir, progress = TRUE)
}
\arguments{
\item{dir}{a directory path within which to find the images}

\item{progress}{logical whether to print progress to the console during processing}
}
\description{
Read all exif data from a directory of image files recursively.
}
