% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exifWrite.R
\name{exifWrite}
\alias{exifWrite}
\title{Apply a new exif tag to image files}
\usage{
exifWrite(files, comment, latlon, coresToUse = TRUE)
}
\arguments{
\item{files}{A vector of files to be modified.}

\item{comment}{A character string to add to the files' exif data as a comment}

\item{latlon}{A numeric vector of length 2, c( latitude, longitude )}

\item{coresToUse}{Either "TRUE" (will use some logic to determine the number of cores to utilise),
or an integer value specifying the number of cores to multi-thread tasks to.}
}
\description{
Apply a new exif tag to image files
}
\keyword{add}
\keyword{comment}
\keyword{exif}
\keyword{image}

