% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileDir.R
\name{fileDir}
\alias{fileDir}
\title{Find the complete directory structure of the folder containing
a specified file.}
\usage{
fileDir(file)
}
\arguments{
\item{file}{character string}
}
\description{
Find the complete directory structure of the folder containing
a specified file.
}
