% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fov.R
\name{fov}
\alias{fov}
\title{calculate the field of view for a given camera setup}
\usage{
fov(focal.length, distance.to.subject.m = 40, focal.mult = 1)
}
\arguments{
\item{focal.length}{numeric vector of focal length values}

\item{distance.to.subject.m}{numeric vector of subject distances in meters}

\item{focal.mult}{numeric vector of focal length multipliers (1.0 is for 35mm format, 1.6 is APS-C, etc.)}
}
\description{
calculate the field of view for a given camera setup
}

