% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColClasses.R
\name{getColClasses}
\alias{getColClasses}
\title{Retrieve the column classes from a data frame.}
\usage{
getColClasses(student)
}
\arguments{
\item{student}{A data frame from which to retrieve classes.}
}
\value{
A vector of classes, represented as character strings.
}
\description{
Retrieve the column classes from a data frame.
}
\keyword{classes}
\keyword{columns,}
\keyword{dataframe,}
\keyword{datatable,}
