% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLatLong.R
\name{getLatLong}
\alias{getLatLong}
\title{Search Google Maps for GPS coordinates, and return as latitude and longitude.}
\usage{
getLatLong(location)
}
\arguments{
\item{location}{A string to pass to Google maps API for searching.}
}
\value{
dataframe
}
\description{
Search Google Maps for GPS coordinates, and return as latitude and longitude.
}
\keyword{GoogleMaps}

