% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPStoEN.R
\name{GPStoEN}
\alias{GPStoEN}
\title{Converting GPS coordinates to Eastings and Northings}
\usage{
GPStoEN(lats, lons, output.projection = "+init=epsg:28355")
}
\arguments{
\item{lats}{numeric vector}

\item{lons}{numeric vector}

\item{output.projection}{character string specifying projection model of outputs}
}
\description{
Converting GPS coordinates to Eastings and Northings
}
\keyword{gps}
\keyword{spatial}
