% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Invert the common function %in%}
\usage{
... \%nin\% NA
}
\arguments{
\item{x}{Value or vector of values to test.}

\item{table}{A vector of values to check against.}
}
\value{
A logical value or vector showing whether each value is NOT represented in the test table.
}
\description{
Invert the common function %in%
}
\keyword{notin}

