% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinRange.R
\name{\%within\%}
\alias{\%within\%}
\title{Also create a wrapper `%within%` for ease of use}
\usage{
x \%within\% range
}
\arguments{
\item{x}{Value or vector of values to test.}

\item{range}{A vector of length at least 2, from which the range will be computed.}
}
\value{
A logical value or vector showing whether each value is within the range.
}
\description{
Also create a wrapper `%within%` for ease of use
}
\keyword{integer}
\keyword{numeric}
\keyword{range}
\keyword{within}

