% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageSlice.R
\name{imageSlice}
\alias{imageSlice}
\title{Take an input image, and slice it into equal parts}
\usage{
imageSlice(image, x.px.out = 1000, y.px.out = 1000, overwrite = FALSE)
}
\arguments{
\item{image}{file path to input image}

\item{x.px.out}{nominal output image dimension, actual dimensions will be calculated}

\item{y.px.out}{nominal output image dimension, actual dimensions will be calculated}

\item{overwrite}{logical whether or not to overwrite old outputs}
}
\description{
Take an input image, and slice it into equal parts
}
