% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate}
\alias{interpolate}
\title{interpolate missing y values to fill a given x series}
\usage{
interpolate(x_in, y_in, x_out)
}
\arguments{
\item{x_in}{input x series}

\item{y_in}{input variable series}

\item{x_out}{output x series}
}
\description{
interpolate missing y values to fill a given x series
}
\keyword{interpolate}
