% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{invDistWInt}
\alias{invDistWInt}
\title{A c++ implementation of inverse distance weighted interpolation}
\usage{
invDistWInt(inputlat, inputlon, inputdata, outputlat, outputlon)
}
\arguments{
\item{inputlat}{vector of input latitudes}

\item{inputlon}{vector of input longitudes}

\item{inputdata}{vector of input data}

\item{outputlat}{vector of output latitudes}

\item{outputlon}{vector of output longitudes}
}
\description{
A c++ implementation of inverse distance weighted interpolation
}
