% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapDistance.R
\name{latDegToKm}
\alias{latDegToKm}
\title{Convert either latitude or longitude distances between angles (degrees) and distances (km)}
\usage{
latDegToKm(degrees)
}
\arguments{
\item{degrees}{Numeric value, degrees for distance}

\item{lat}{Numeric value, latitude position}

\item{km}{Numeric value, distance in km}
}
\value{
numeric value
}
\description{
Convert either latitude or longitude distances between angles (degrees) and distances (km)
}
\keyword{gis}
\keyword{maps}

