% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchColClasses.R
\name{matchColClasses}
\alias{matchColClasses}
\title{Match the column classes of one data frame to another. Useful before calling `rbind`.
Note: this is just an easy way to use `getColClasses` and `applyColClasses` in a single step.}
\usage{
matchColClasses(df1, df2)
}
\arguments{
\item{df1}{A data frame or data table to be changed.}

\item{df2}{A data frame or data table to be matched. This one will be unchanged}
}
\value{
A data frame, with column classes adjusted as necessary.
}
\description{
Match the column classes of one data frame to another. Useful before calling `rbind`.
Note: this is just an easy way to use `getColClasses` and `applyColClasses` in a single step.
}
\keyword{classes}
\keyword{columns,}
\keyword{dataframe,}
\keyword{datatable,}

