% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newRow.R
\name{newRow}
\alias{newRow}
\title{A function for adding a data row to a given data frame}
\usage{
newRow(x)
}
\arguments{
\item{x}{a data.frame or data.table to have a row added}
}
\description{
A function for adding a data row to a given data frame
}
