% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgLoad.R
\name{pkgLoad}
\alias{pkgLoad}
\title{activate the "install.packages" function, only if the package is not already installed,
then load that package}
\usage{
pkgLoad(packages)
}
\arguments{
\item{packages}{A vector of packages to be checked or installed}
}
\description{
activate the "install.packages" function, only if the package is not already installed,
then load that package
}
\keyword{packages}

