% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_expand.R
\name{range_expand}
\alias{range_expand}
\title{A function for use when adding a data row to a given data frame}
\usage{
range_expand(x, expand.lims = 0, shift = 0, na.rm = TRUE)
}
\arguments{
\item{x}{a vector of values}

\item{expand.lims}{how much to expand numeric value or length 2 vector}

\item{shift}{optional shift parameter}

\item{na.rm}{as per base range function}
}
\description{
A function for use when adding a data row to a given data frame
}
