% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMean.R
\name{runMean}
\alias{runMean}
\title{Smooth a numeric vector using moving window averaging.}
\usage{
runMean(input, window, overlap = FALSE, fill.NAs = TRUE, na.Rm = TRUE)
}
\arguments{
\item{input}{A numeric vector of data to be smoothed.}

\item{window}{Numeric value for the size of each "group" to be averaged.}

\item{overlap}{Logical. Should the windows overlap (by half the window size)?}

\item{fill.NAs}{Logical. If TRUE, only the centre point of each window is filled.
Goes to TRUE if `overlap` is TRUE.}

\item{na.Rm}{Logical, to be passed to `mean` function as `na.rm`.}
}
\description{
Smooth a numeric vector using moving window averaging.
}
\keyword{mean}
\keyword{rolling}
\keyword{smooth}
\keyword{window}

