% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentImage.R
\name{segmentImage}
\alias{segmentImage}
\title{segmentImage}
\usage{
segmentImage(input.image, output.grid = NULL)
}
\arguments{
\item{input.image}{file name of input jpeg}

\item{output.grid}{integer vector of format "c( x, y )"}
}
\description{
Break up a .jpg image into segments, before outputting to individual files
}

