% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_.R
\name{seq_across}
\alias{seq_across}
\title{Return a sequential vector along x column-wise.}
\usage{
seq_across(x)
}
\arguments{
\item{x}{data.frame (incl. data.table) or matrix object}
}
\value{
vector sequence
}
\description{
Return a sequential vector along x column-wise.
}
\keyword{data.frame}
\keyword{data.table}
\keyword{matrix}
