% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_.R
\name{seq_dim}
\alias{seq_dim}
\title{Return a sequential vector along x in the dimension specified.}
\usage{
seq_dim(x, mar)
}
\arguments{
\item{x}{data.frame (incl. data.table) or array (incl. matrix) object}

\item{mar}{integer for margin along which to create the sequence}
}
\value{
vector sequence
}
\description{
Return a sequential vector along x in the dimension specified.
}
\keyword{array}
\keyword{data.frame}
\keyword{data.table}
\keyword{matrix}
