% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift.R
\name{shiftLag}
\alias{shiftLag}
\title{Shift values of a vector by a given amount.}
\usage{
shiftLag(input, n = 1L, fill = NA)
}
\arguments{
\item{input}{A vector to be shifted.}

\item{n}{An integer value, how far to shift.}

\item{fill}{A value to fill in empty cells created during the shift operation.}
}
\value{
A vector the same length as the input vector, but with values shifted.
}
\description{
Shift values of a vector by a given amount.
}
\keyword{lag,}
\keyword{lead}
\keyword{shift,}
\keyword{vector,}
