% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceFolder.R
\name{sourceFolder}
\alias{sourceFolder}
\title{Source all files in a particular folder.}
\usage{
sourceFolder(folder, except = NULL, use.regex = FALSE, quietly = FALSE)
}
\arguments{
\item{folder}{Character string directory link to folder.}

\item{except}{Character vector, files to avoid sourcing}

\item{use.regex}{When eliminating `except` files, should we use regular expression search?}

\item{quietly}{Logical, should we output running status along the way?}
}
\value{
invisible character vector of files sourced
}
\description{
Source all files in a particular folder.
}
\keyword{function,}
\keyword{load,}
\keyword{script}
\keyword{source,}

