% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringLength.R
\name{stringLength}
\alias{stringLength}
\title{Change the length of a character string, placing the input string at
one end of the output string (if lengthening), or taking one end of
the input string only (when shortening). Useful to match lengths of
character strings, where inputs vary in length.}
\usage{
stringLength(input, length.out, which = c("head", "tail"), fill = NA)
}
\arguments{
\item{input}{A character vector (coerced to character if not already)}

\item{length.out}{Integer or numerical value for output character length}

\item{which}{Options "head" (start of string) or "tail" (end of string)}

\item{fill}{Character vector of any length, used to fill in missing
characters where input needs to be lengthened. NA (default) will not
lengthen, meaning output strings may be shorter than "length.out".}
}
\description{
Change the length of a character string, placing the input string at
one end of the output string (if lengthening), or taking one end of
the input string only (when shortening). Useful to match lengths of
character strings, where inputs vary in length.
}
\keyword{character}
\keyword{length}
\keyword{string}

