% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripWhiteSpace.R
\name{stripWhiteSpace}
\alias{stripWhiteSpace}
\title{Strip the white space from around a character string}
\usage{
stripWhiteSpace(input, which = "both")
}
\arguments{
\item{input}{A character vector from which white space is to be stripped.}

\item{which}{One of "both", "leading", or "trailing". Specifying from which
end (or both) to strip white space.}
}
\description{
Strip the white space from around a character string
}
\keyword{character}
\keyword{leading}
\keyword{string}
\keyword{trailing}
\keyword{whitespace}
