% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supeRbind.R
\name{supeRbind}
\alias{supeRbind}
\title{supeRbind
Match the column classes of one data frame to another. Useful before calling `rbind`.
Note: this is just an easy way to use `getColClasses` and `applyColClasses` in a single step.}
\usage{
supeRbind(master, student)
}
\arguments{
\item{master}{A data frame or data table acting as MASTER.}

\item{student}{A data frame or data table, will "learn" from the master.}
}
\value{
A data frame, with column classes adjusted as necessary.
}
\description{
supeRbind
Match the column classes of one data frame to another. Useful before calling `rbind`.
Note: this is just an easy way to use `getColClasses` and `applyColClasses` in a single step.
}
\keyword{classes}
\keyword{columns,}
\keyword{dataframe,}
\keyword{datatable,}

