% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringLength.R
\name{tagsLongToShort}
\alias{tagsLongToShort}
\title{tagsLongToShort
Specific function for converting hex PIT tag readings to a 6 character
base 10 integer, as used by the older "Automated Penguin Monitoring System".}
\usage{
tagsLongToShort(tags)
}
\arguments{
\item{tags}{A character vector (coerced to character if not already)}
}
\value{
A vector of character strings, the same vector length as the input one.
}
\description{
tagsLongToShort
Specific function for converting hex PIT tag readings to a 6 character
base 10 integer, as used by the older "Automated Penguin Monitoring System".
}
\keyword{character}
\keyword{conversion}
\keyword{hex}
\keyword{integer}
\keyword{length}
\keyword{string}

