% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testDF.R
\name{testDF}
\alias{testDF}
\title{testDF, a function for creating a data frame for mucking around with.}
\usage{
testDF(rows, DT = FALSE)
}
\arguments{
\item{rows}{how big should the data frame be}

\item{DT}{should the output be a data.table? (data.frame is output if FALSE)}
}
\description{
testDF, a function for creating a data frame for mucking around with.
}

