% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichComputer.R
\name{whichComputer}
\alias{whichComputer}
\title{A function to find the computer currently in use}
\usage{
whichComputer()
}
\description{
A function to find the computer currently in use
}
\keyword{OperatingSystem}
\keyword{UserName}
\keyword{computer}

