% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/withinRange.R
\name{withinRange}
\alias{withinRange}
\title{Test whether one or more values fall within a specified range.}
\usage{
withinRange(x, range, inclusive = TRUE)
}
\arguments{
\item{x}{Value or vector of values to test.}

\item{range}{A vector of length at least 2, from which the range will be computed.}

\item{inclusive}{Logical value, whether to include the bound of the range.}
}
\value{
A logical value or vector showing whether each value is within the range.
}
\description{
Test whether one or more values fall within a specified range.
}
\keyword{integer}
\keyword{numeric}
\keyword{range}
\keyword{within}
