% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyColClasses.R
\name{applyColClasses}
\alias{applyColClasses}
\title{Apply column classes to a data frame. Useful before calling `rbind`.}
\usage{
applyColClasses(df, col.classes)
}
\arguments{
\item{col.classes}{A vector of classes to be applied. Probably an output from `getColClasses`.}
}
\value{
A data frame, with column classes adjusted as necessary.
}
\description{
Apply column classes to a data frame. Useful before calling `rbind`.
}
\keyword{classes}
\keyword{columns,}
\keyword{dataframe,}
\keyword{datatable,}

