% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exifRead.R
\name{exifRead}
\alias{exifRead}
\title{Extract filename, directory, date, and time from image EXIF data.}
\usage{
exifRead(files, coresToUse = TRUE)
}
\arguments{
\item{files}{A vector of files to be analysed.}

\item{coresToUse}{Either "TRUE" (will use some logic to determine the number of cores to utilise),
or an integer value specifying the number of cores to multi-thread tasks to.}
}
\description{
Extract filename, directory, date, and time from image EXIF data.
}
\keyword{date}
\keyword{exif}
\keyword{image}
\keyword{name}
\keyword{time}
