% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColClasses.R
\name{getColClasses}
\alias{getColClasses}
\title{Retrieve the column classes from a data frame.}
\usage{
getColClasses(df)
}
\description{
Retrieve the column classes from a data frame.
}
\keyword{classes}
\keyword{columns,}
\keyword{dataframe,}
\keyword{datatable,}

