% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageContrast.R
\name{imageContrast}
\alias{imageContrast}
\title{calculate and apply an S curve to modify the contrast of an image}
\usage{
imageContrast(input, k = 0.3)
}
\arguments{
\item{input}{A file extension to a jpg file, or an array imported from an image file.}

\item{k}{Value between -1 and 1. Positive to increase contrast, negative to decrease.}
}
\description{
calculate and apply an S curve to modify the contrast of an image
}
\keyword{contrast}
\keyword{curve}
\keyword{image}
\keyword{modify}

