% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchColNames.R
\name{matchColNames}
\alias{matchColNames}
\title{Match the column names between 2 data frames. Useful to use before doing calling `rbind`.}
\usage{
matchColNames(df1, df2)
}
\arguments{
\item{df1}{A dataframe or datatable, to be changed to match columns of df2.
df2 A "master" dataframe or datatable, df1 will be changed to match this.}
}
\description{
Match the column names between 2 data frames. Useful to use before doing calling `rbind`.
}
\keyword{columns,}
\keyword{dataframe,}
\keyword{datatable,}
\keyword{matching}

