% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentImage.R
\name{segmentImage}
\alias{segmentImage}
\title{segmentImage}
\usage{
segmentImage(input.image, output.grid = NULL, output.px = NULL)
}
\arguments{
\item{input.image}{file name of input jpeg}

\item{output.grid}{integer vector specifying how to divide the input image. format "c( x, y )"}

\item{output.px}{integer vector specifying nominal dimensions of output images. format "c( x, y )"}
}
\description{
Break up a .jpg image into segments, before outputting to individual files
}

